requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.23.2
picamera2
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]
python-multipart
gpiozero
RPi.GPIO
